Update versioning & config improvements! - Updated 'update-version.ts' to use metadata instead of cli.ts for version handling - Improved some config commands across multiple files - Adjusted tests for better performance and accuracy - Refactored a few core modules for clarity and maintainability - Removed outdated 'src/prompts/index.ts' file - Added or modified functions/vars in several files for enhanced functionality
