+++ src/cli/commands/config/remove.ts
@@ -1,8 +1,8 @@
 import { Command } from 'commander';
+import fsExtra from 'fs-extra';
 import { ConfigManager } from '../../../core/config';
 import { Logger } from '../../../utils/logger';
 import { getConfigFileInfo } from '../../utils/get-friendly-source';
-import fsExtra from 'fs-extra';
 
 export const registerRemoveCommands = (configCommand: Command) => {
   const removeCommand = configCommand.command('remove').description('Remove configuration files');
