--- src/prompts/index.ts
@@ -1,5 +1,0 @@
-export { PROMPTS } from '../constants/prompts';
-export { defaultPrompt } from './default';
-export { conventionalPrompt } from './conventional';
-export { simplePrompt } from './simple';
-export { detailedPrompt } from './detailed';
