+++ test/setup.ts
@@ -6,7 +6,7 @@ import * as os from 'os';
 export const mockFs = {
   files: new Map<string, string>(),
   dirs: new Set<string>(),
-  
+
   async mkdir(dir: string, options?: { recursive: boolean }) {
     this.dirs.add(dir);
     if (options?.recursive) {
@@ -18,25 +18,25 @@ export const mockFs = {
       }
     }
   },
-  
+
   async writeFile(filePath: string, content: string) {
     this.files.set(filePath, content);
     // Ensure parent directory exists
     const dir = path.dirname(filePath);
     this.dirs.add(dir);
   },
-  
+
   async readFile(filePath: string, encoding: string) {
     return this.files.get(filePath) || '';
   },
-  
+
   async access(filePath: string) {
     if (this.files.has(filePath) || this.dirs.has(filePath)) {
       return;
     }
     throw new Error('File not found');
   },
-  
+
   async rm(dir: string, options?: { recursive: boolean; force: boolean }) {
     if (options?.recursive) {
       for (const file of this.files.keys()) {
@@ -50,7 +50,7 @@ export const mockFs = {
         }
       }
     }
-  }
+  },
 };
 
 // Mock Git commands
@@ -65,7 +65,7 @@ export const mockGit = {
       }
     }
     return '';
-  }
+  },
 };
 
 // Mock Ollama API responses
@@ -73,43 +73,43 @@ export const mockOllamaResponses = {
   models: {
     models: [
       { name: 'mistral:7b-instruct', modified_at: '2024-03-20T12:00:00Z' },
-      { name: 'llama2:7b', modified_at: '2024-03-20T12:00:00Z' }
-    ]
+      { name: 'llama2:7b', modified_at: '2024-03-20T12:00:00Z' },
+    ],
   },
   generate: {
     model: 'mistral:7b-instruct',
     response: 'feat: add new feature for improved performance',
-    done: true
+    done: true,
   },
   health: {
-    status: 'ok'
-  }
+    status: 'ok',
+  },
 };
 
 // Mock fetch for Ollama API calls
 export function mockOllamaFetch() {
   const originalFetch = global.fetch;
-  
+
   global.fetch = async (input: RequestInfo | URL, init?: RequestInit) => {
     const url = input instanceof URL ? input : new URL(input.toString());
-    
+
     // Mock different endpoints
     if (url.pathname === '/api/tags') {
       return new Response(JSON.stringify(mockOllamaResponses.models));
     }
-    
+
     if (url.pathname === '/api/generate') {
       return new Response(JSON.stringify(mockOllamaResponses.generate));
     }
-    
+
     if (url.pathname === '/api/health') {
       return new Response(JSON.stringify(mockOllamaResponses.health));
     }
-    
+
     // Default to original fetch for other requests
     return originalFetch(input, init);
   };
-  
+
   return () => {
     global.fetch = originalFetch;
   };
@@ -135,7 +135,11 @@ export const mockConfig = {
   promptTemplate: 'default',
 };
 
-// Helper function to create mock git diff
+/**
+ * Creates a mock Git diff string for testing purposes.
+ *
+ * @returns A string containing a mock Git diff output for a new file
+ */
 export function createMockGitDiff() {
   return `diff --git a/test.txt b/test.txt
 new file mode 100644
@@ -151,9 +155,12 @@ index 0000000..1234567
 beforeAll(async () => {
   // Setup Ollama API mocking
   mockOllamaFetch();
-  
+
   // Create mock files
-  await mockFs.writeFile(mockConfig.promptFile, 'Generate a commit message for the following changes:\n{{diff}}');
+  await mockFs.writeFile(
+    mockConfig.promptFile,
+    'Generate a commit message for the following changes:\n{{diff}}',
+  );
   await mockFs.writeFile(mockConfig.configFile, JSON.stringify(mockConfig, null, 2));
 });
